feat(edificios): fusión 11 fuentes, 12 arquetipos vascos completos, tejados AAA
FusionadorEdificiosUltra.cs — reescritura completa con prioridades reales:
  - Altura: LIDAR > DSM (.asc grid) > catastro > OSM height > levels×3.2m
  - Color fachada: mapillary 8m > OSM building:colour > ortofoto ±8% Perlin
  - Color tejado: ortofoto centroide > OSM roof:colour > tipo/época
  - Enriquecimiento de buildings_osm_rico (amenity, shop, sport, tags)
  - ArquetipoVasco enum con 12 valores, GetArquetipoConAnio() con año real
  - API long id + legacy compat int id

SistemaEdificiosAAA.cs — 12 arquetipos vascos procedurales:
  a. UrbanoPre1940: arenisca #c8a882, balcones forja, tejado 32°, chimeneax2
  b. Bloque1940-1975: ladrillo #8b4513, ventanas rectas, balcón hormigón
  c. ModernoPost1975: revoco blanco #f5f0e8, tejado plano, cajas AC
  d. Caserío: cal #faf7f0, viga madera 1.2m vuelo, tejado 42°
  e. Bar: azulejo PB, rótulo OSM, canopy, barril, luz 2700K
  f. Comercio: escaparate 80% transparente, persiana, rótulo
  g. Iglesia: piedra oscura, campanil, contrafuertes, atrio cobblestone
  h. NaveIndustrial: chapa metálica R=0.8, puerta corredera 3×4m, lucernario
  i. EquipamientoPublico: ladrillo institucional, marquesina, megáfono
  j. Frontón: hormigón blanco, marcas impacto procedurales, red metálica
  k. Aparcamiento: cubierto losa+pilares / descubierto líneas blancas
  l. Solar: escombros bicolor + graffiti vasco
  LOD 4 niveles (<40/<120/<350/<700m), GPU Instancing + MaterialPropertyBlock
  ValidarEdificios(): altura vs LIDAR, base vs terrain, rótulos, estadísticas

GeneradorTejadosAAA.cs — nuevo script de tejados:
  Forma: OSM roof:shape > tipo+niveles > arquetipo
  Gabled 32°/42°/52°/25°/12°, Hipped, Flat con pretil 0.35m
  Color: sample ortofoto → terracota #b5451b vs pizarra #5a6474
  Chimney_Stone (residencial) + Chimney_Pipe (moderno)
  Canalón + bajante metal gris oscuro cada 8m

GeneradorGeometriaPrecisa.cs — mejoras:
  Perturbación Perlin vértices σ=0.015m
  Zócalo 0.28m más oscuro perimetral
  Cornisa voladizo 0.12-0.18m con perfil moldurado
  Vierteaguas (canopy_side) bajo cada ventana (pisos >0)
  Altura long(id) para compatibilidad con FusionadorEdificiosUltra

GeneradorFachadasAAA.cs — reescritura:
  Grid: columnas = FloorToInt(ancho/1.15f), Kit_Window_Upper_Convex pisos 0-1
  Planta baja: puerta+Canopy_Full residencial, escaparate 80% comercio
  Balcones pre-1940: Canopy_Full + Canopy_Beam_Short + barandilla hierro negro
  Color: sample ortofoto centroide 2m altura → tint ±8% Perlin via MPB
  Panel eléctrico fachada trasera comercios altura 1.4m
  Contenedores basura procedurales cada 3-4 portales

+++ Assets/Scripts/SistemaEdificiosAAA.cs
@@ -139,7 +139,7 @@ public class SistemaEdificiosAAA : MonoBehaviour
         // Consultar altura LIDAR si disponible
         float altura = FusionadorEdificiosUltra.Instance != null
             ? Mathf.Max(GeoDataAlsasua.ALT_PLANTA,
-                FusionadorEdificiosUltra.Instance.GetAlturaOptima(e.id, altOSM))
+                FusionadorEdificiosUltra.Instance.GetAlturaOptima((long)e.id, altOSM))
             : altOSM;
 
         var mesh = GenerarMeshEdificio(verts2D, suelo, altura);
@@ -147,7 +147,7 @@ public class SistemaEdificiosAAA : MonoBehaviour
 
         // Determinar arquetipo
         ArquetipoVasco arquetipo = FusionadorEdificiosUltra.Instance != null
-            ? FusionadorEdificiosUltra.Instance.GetArquetipoConAnio(e.id, e.type, e.levels)
+            ? FusionadorEdificiosUltra.Instance.GetArquetipoConAnio((long)e.id, e.type, e.levels)
             : FusionadorEdificiosUltra.ArquetipoDesdeOSM(e.type, e.levels, "", "", "");
 
         string nombre = string.IsNullOrEmpty(e.name)
@@ -160,12 +160,12 @@ public class SistemaEdificiosAAA : MonoBehaviour
         var mf = go.AddComponent<MeshFilter>();
         mf.sharedMesh = mesh;
 
-        Material matFachada = MaterialFachadaPorArquetipo(arquetipo, e.id);
+        Material matFachada = MaterialFachadaPorArquetipo(arquetipo, (long)e.id);
 
         // Aplicar variación de color por MaterialPropertyBlock (sin instanciar material)
         var mr = go.AddComponent<MeshRenderer>();
         mr.sharedMaterial = matFachada;
-        AplicarVariacionColor(mr, arquetipo, e.id);
+        AplicarVariacionColor(mr, arquetipo, (long)e.id);
         mr.shadowCastingMode = UnityEngine.Rendering.ShadowCastingMode.On;
 
         // Collider en radio < 200m del origen
